requirements.txt
numpy>=1.24
torch>=2.1
tqdm>=4.64
pyyaml>=6.0
matplotlib>=3.7

